backend/requirements.txt
gunicorn==19.9.0
Django==4.1
djangorestframework==3.13.1
django-filter==22.1
djangorestframework-simplejwt==5.2.0
python-dotenv==0.21.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
django-multiselectfield==0.1.12
Faker==14.2.0
django-mock-queries==2.1.7
pydocstyle==6.1.1
flake8==5.0.4
django-nose==1.4.7
coverage==6.4.4
pytest-django==4.5.2
pytest-env==0.6.2
pytest-dependency==0.5.1
pytest-cov==2.7.1
pytest==7.1.3
coverage==6.4.4
drf_yasg==1.21.3
requests==2.28.1